Implement factory pattern to create parsers
After discussions with Steinar, I've reimplemented the way that the
parsers are initiated. Previously I had implemented a singleton pattern
in order to ensure that only a single instance of each parser class was
created, but now I've made the parser implementations more generic and
handled the parser initiation and signing up the subscribers via the qc
engine instead. I think that this makes the code easier to reason about.