Build on readthedocs.org is broken
This hopefully fixes the readthedocs build. I followed their
instructions on how to setup a new project, and noticed that the lines
that were removed here were the ones which were not in the example
config. I've tried generating the docs locally and it looks alright, so
I'm hoping this will fix the trouble on readthedocs...