Remove UnidentifiedIndexHandler from config
We want to release the changes to the UndeterminedPercentageHandler, but
the UnidentifiedIndexHandler is not yet ready for production. Removing
the latter from the config means that we can release and keep working on
checking the indexes until that is ready. Essentially this means that we
have dead code in the production release (but that should be fine as
this module should not interfer at all with the other modules).