UnidentifiedIndexHandler not enabled
The UnidentifiedIndexHandler will now throw an not implemented error if
it is instantiated, so that no-one trys to add it into their config
eventhough it is not actually done yet.

This is a somewhat ugly work-around to spare us from having to
cherry-pick the parts of the code required to have multiple parsers for
a handler.